+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add `Command.parse` function
 
 ## 2.0.0 - 2020-05-13
 - [**BC**] Change all DTO structures to PascalCase, and let the serialization dealt with a format